Changed length data types from uint8_t to size_t

+++ include/of2g.h
@@ -3,7 +3,7 @@
 
 #include <inttypes.h>
 #include <stdbool.h>
-
+#include <stddef.h>
 // Need to confirm what the actual max size for an OF2G frame is
 // it should be less than 256 bytes though, since the He frames are
 // max 256 bytes including their metadata (from what I remember anyway)
@@ -57,7 +57,7 @@ uint8_t of2g_get_data_content(of2g_frame_t * frame, uint8_t * out);
 //
 // This function should return true if the frame is built successfully, and
 // should return false if the frame can't be built for any reason.
-bool of2g_build_data_frame(unsigned char * buffer, uint8_t length, unsigned char fid, of2g_frame_t * out);
+bool of2g_build_data_frame(unsigned char * buffer, size_t length, unsigned char fid, of2g_frame_t * out);
 
 // This function should create an OF2G ack frame to be used as a response for
 // `data_frame`, storing it in `out`. It should return true as long as the frame
